# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Vault API
  version: 0.1.0
  description: |
    Vault HTTP JSON API.

    Conventions:
    - JSON request/response bodies use `application/json` unless otherwise noted.
    - Error responses use the `ErrorResponse` schema.
    - Pagination: list endpoints use `limit` with cursor-based pagination via `cursor`/`nextCursor`. The media list also supports page-based pagination via `page`/`nextPage` when `cursor` is omitted. Max `limit` for media list is 100.
    - IDs are strings; media IDs are validated as UUIDs in path parameters.
    - Timestamps are RFC3339 `date-time` strings.
    - Auth: `requireAuth` prefers `Authorization: Bearer <token>` and falls back to the `access_token` cookie. `/api/auth/me` is cookie-only.
    - Query array params use `style: form` with `explode: true` (e.g., `tags=one&tags=two`).
servers:
  - url: http://127.0.0.1:8000
    description: Local dev
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/bundles:
    get:
      summary: List bundles for the current user
      tags:
        - Bundles
      operationId: bundles_list
      responses:
        "200":
          description: Bundle list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleListResponse"
              examples:
                success:
                  value:
                    bundles:
                      - id: "c1d2e3f4-0001-0000-0000-000000000001"
                        name: "Tax documents 2025"
                        description: "W2, receipts, and filings"
                        starred: true
                        itemCount: 7
                        coverMediaId: "7a1b54a0-5e98-4f1e-a3f5-2b2b3e9f08c9"
                        createdAt: "2026-01-01T00:00:00Z"
                        updatedAt: "2026-03-15T12:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a bundle
      tags:
        - Bundles
      operationId: bundles_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleCreateRequest"
            examples:
              create:
                value:
                  name: "Tax documents 2025"
                  description: "W2, receipts, and filings"
      responses:
        "201":
          description: Bundle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleSummaryResponse"
              examples:
                success:
                  value:
                    bundle:
                      id: "c1d2e3f4-0001-0000-0000-000000000001"
                      name: "Tax documents 2025"
                      description: "W2, receipts, and filings"
                      starred: false
                      itemCount: 0
                      coverMediaId: null
                      createdAt: "2026-03-17T00:00:00Z"
                      updatedAt: "2026-03-17T00:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/bundles/{id}:
    get:
      summary: Get bundle detail with items
      tags:
        - Bundles
      operationId: bundles_get
      parameters:
        - $ref: "#/components/parameters/BundleId"
      responses:
        "200":
          description: Bundle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleDetailResponse"
              examples:
                success:
                  value:
                    bundle:
                      id: "c1d2e3f4-0001-0000-0000-000000000001"
                      name: "Tax documents 2025"
                      description: "W2, receipts, and filings"
                      starred: true
                      itemCount: 2
                      coverMediaId: "7a1b54a0-5e98-4f1e-a3f5-2b2b3e9f08c9"
                      createdAt: "2026-01-01T00:00:00Z"
                      updatedAt: "2026-03-15T12:00:00Z"
                      items:
                        - mediaId: "7a1b54a0-5e98-4f1e-a3f5-2b2b3e9f08c9"
                          order: 0
                          addedAt: "2026-01-01T00:00:00Z"
                          title: "W2 2025"
                          mimeType: "application/pdf"
                          thumbState: "READY"
                          thumbnailKey: "usr_123/7a1b54a0/thumb.webp"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update bundle name or description
      tags:
        - Bundles
      operationId: bundles_update
      parameters:
        - $ref: "#/components/parameters/BundleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundlePatchRequest"
            examples:
              update:
                value:
                  name: "Tax docs 2025 (updated)"
      responses:
        "200":
          description: Bundle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              examples:
                success:
                  value:
                    ok: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a bundle
      description: Deletes the bundle record. Media items are not deleted.
      tags:
        - Bundles
      operationId: bundles_delete
      parameters:
        - $ref: "#/components/parameters/BundleId"
      responses:
        "204":
          description: Bundle deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/bundles/{id}/items:
    post:
      summary: Add media items to a bundle
      tags:
        - Bundles
      operationId: bundles_addItems
      parameters:
        - $ref: "#/components/parameters/BundleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleAddItemsRequest"
            examples:
              add:
                value:
                  mediaIds:
                    - "7a1b54a0-5e98-4f1e-a3f5-2b2b3e9f08c9"
                    - "8f0c1c5c-1c60-4c06-9e3d-7d0bfae7a4da"
      responses:
        "200":
          description: Items added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              examples:
                success:
                  value:
                    ok: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/bundles/{id}/items/{mediaId}:
    delete:
      summary: Remove a media item from a bundle
      tags:
        - Bundles
      operationId: bundles_removeItem
      parameters:
        - $ref: "#/components/parameters/BundleId"
        - in: path
          name: mediaId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Item removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/bundles/{id}/items/order:
    put:
      summary: Reorder items within a bundle
      tags:
        - Bundles
      operationId: bundles_reorderItems
      parameters:
        - $ref: "#/components/parameters/BundleId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleOrderRequest"
            examples:
              reorder:
                value:
                  orderedIds:
                    - "8f0c1c5c-1c60-4c06-9e3d-7d0bfae7a4da"
                    - "7a1b54a0-5e98-4f1e-a3f5-2b2b3e9f08c9"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
              examples:
                success:
                  value:
                    ok: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/bundles/{id}/star:
    post:
      summary: Toggle the starred state of a bundle
      tags:
        - Bundles
      operationId: bundles_toggleStar
      parameters:
        - $ref: "#/components/parameters/BundleId"
      responses:
        "200":
          description: Starred state toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleStarResponse"
              examples:
                starred:
                  value:
                    ok: true
                    starred: true
                unstarred:
                  value:
                    ok: true
                    starred: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Reminders
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token

  # ---------------------------------------------------------------------------
  # Reusable parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    BundleId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Bundle ID.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            badRequest:
              $ref: "#/components/examples/ErrorBadRequest"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              $ref: "#/components/examples/ErrorUnauthorized"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            notFound:
              $ref: "#/components/examples/ErrorNotFound"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            serverError:
              $ref: "#/components/examples/ErrorInternalServer"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
    ErrorResponse:
      type: object
      description: Error response payload from Fastify and rate limiting.
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
        limit:
          type: integer
        windowMs:
          type: integer
    OkResponse:
      type: object
      description: Generic acknowledgement response.
      properties:
        ok:
          type: boolean
      required:
        - ok

    # --- Auth ---
...
    MediaWorkerState:
      type: string
      description: Background processing state.
      enum:
        - PENDING
        - READY
        - ERROR
        - FAILED
...
    BundleSummary:
      type: object
      description: Bundle summary returned in list responses.
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: [string, "null"]
        starred:
          type: boolean
        itemCount:
          type: integer
        coverMediaId:
          type: [string, "null"]
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - starred
        - itemCount
        - createdAt
        - updatedAt
    BundleItem:
      type: object
      description: A media item within a bundle.
      properties:
        mediaId:
          type: string
          format: uuid
        order:
          type: integer
        addedAt:
          type: string
          format: date-time
        title:
          type: [string, "null"]
        mimeType:
          type: string
        thumbState:
          $ref: "#/components/schemas/MediaWorkerState"
        thumbnailKey:
          type: [string, "null"]
      required:
        - mediaId
        - order
        - addedAt
        - mimeType
        - thumbState
    BundleDetail:
      description: Bundle summary with its ordered item list.
      allOf:
        - $ref: "#/components/schemas/BundleSummary"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/BundleItem"
          required:
            - items
    BundleListResponse:
      type: object
      description: Bundle list response envelope.
      properties:
        bundles:
          type: array
          items:
            $ref: "#/components/schemas/BundleSummary"
      required:
        - bundles
    BundleSummaryResponse:
      type: object
      description: Single bundle summary response envelope.
      properties:
        bundle:
          $ref: "#/components/schemas/BundleSummary"
      required:
        - bundle
    BundleDetailResponse:
      type: object
      description: Bundle detail response envelope.
      properties:
        bundle:
          $ref: "#/components/schemas/BundleDetail"
      required:
        - bundle
    BundleCreateRequest:
      type: object
      description: Request to create a new bundle.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 1000
      required:
        - name
    BundlePatchRequest:
      type: object
      description: Update bundle name or description (at least one field required).
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, "null"]
          maxLength: 1000
    BundleAddItemsRequest:
      type: object
      description: Add media items to a bundle (1–100 IDs).
      properties:
        mediaIds:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            format: uuid
      required:
        - mediaIds
    BundleOrderRequest:
      type: object
      description: Reorder items in a bundle by providing the full ordered list of media IDs.
      properties:
        orderedIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
      required:
        - orderedIds
    BundleStarResponse:
      type: object
      description: Result of toggling the starred state.
      properties:
        ok:
          type: boolean
        starred:
          type: boolean
      required:
        - ok
        - starred

    # --- Reminders ---
...
  examples:
    ErrorUnauthorized:
      value:
        statusCode: 401
        error: "Unauthorized"
        message: "Unauthorized"
    ErrorBadRequest:
      value:
        statusCode: 400
        error: "Bad Request"
        message: "Bad Request"
    ErrorNotFound:
      value:
        statusCode: 404
        error: "Not Found"
        message: "Not Found"
...
    ErrorInternalServer:
      value:
        statusCode: 500
        error: "Internal Server Error"
        message: "Internal Server Error"